move modals reducers to the store dir
