+++ src/store/modals/index.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
 
-import channelDeletion from './channelDeletion/channelDeletion';
-import channelEditing from './channelEditing/channelEditing';
-import alert from './alert/alert';
+import channelDeletion from './channelDeletion';
+import channelEditing from './channelEditing';
+import alert from './alert';
 
 export default combineReducers({
   alert: alert.reducer,
